Basic Table Operations Input Tsv Parser TreeTableView Adduct View Next Button RT in Minutes Fixed EIC Open multiple mzXML Color! Width! Default and set color, width Comments Simple Progress Bar / Threading Started on Batch Tab Batch Tabs Renaming Simple peak picking xAxis correction (all adducts have same min max) More peak picking Correlation more peak picking... ... no more peak picking, comparison of AVGEIC generation of Interpolated EICs Quality of peak in the middle Visualization of ideal peak Method restructuring Backup before Datastructure rework rework... Matrix for shortest path Optimized Path calculation Path visualization, Peak Prop Vis Toggle shiftview Fixes, load mzxml context menu, focus on clicked adduct in adductview (needs fixes) Before Dataset-Maps After Dataset-Maps Acitve Dataset fixed currentdataset not being initialized changed/fixed POM. netbeans has problems with maven+javafx, should work now. Follow-up problems not foreseeable. Wavelet peak picking M/Z chart shows ppm M/Z shift calculation/cleaning per file Fixed ChartGenerator Loops Score for selected Files Delete Files Delete Batches minimized RAM fixed Parsing fixed File Deletion Deactivate Files Change Properties for selected files Highlight selected files in Adductview, select new ones Change color, width of files changes adductview/shift. Unselect in shiftview Fixed Adductview, no selection, selection update Hover, shiftview Open Adduct from shiftview Fixed Listener memomryleak scroll to selected OGroup when double click in Shiftview, default sort by RT Accordion rework... ...done, draw multiple Datasets, select across Datasets included RT-range for shift calculation, commented out Active Dataset De/activate whole Dataset Parameters additional scores Naive peak preparation Naive peak picking and fitting, output to console Output to file, additional scores in Table Proper sorting and start on Adductview labels fixed additional scores, min sorting for Scorepeakclose select Opacity Mode in Shiftview Experimental Distance Opacity Mode No need for PropArray Change PeakPicking and Baseline on the fly Smooth PropArray Fixed missing M/Z values/chart Selection Rectangle in Shiftview Set Penalty Area Progress bar for Adductview Disable Charts in Adductview Show peak borders Fix Output, fix Adductview, fix peak borders fixed issue with RT Tolerance>2 small changes Visualize Penalty Area Started with Peak matrix Visualization Moved peak vis, small fixes Implemented certainty small fixes, fleshed out certainty show avg. certainty and %peaks found in Batch show files in peak matrix vis started on peak Mode peak Mode implemented, only visual peak Mode manual peak, only visual peak Mode manual peak, peak picking addPeak Button fixed manual peak picking included cutoff in manual peak picking, only look for peak in inner 70% range changed peak marker, don't delete manual peaks during automatic peak picking, don't add manual peak if peak is already there delete peaks small fix more adducts delete empty slices, double->float don't generate duplicate adducts fix selection in adductview small fixes fixed inactive file selection in shiftview bytes for MZ values, only 25% memory needed fix (delete Bins) started on IO working IO, needs tuning Shiftview own thread fixed peak view, more work in IO start with MemoryMappedFile MemoryMappedFile manual peak pick out of range fix, start new IOThread when crashed Output includes generated Adducts Adduct Parameters various GUI changes various changes tried to fix binning added start end RT value started on rework(no interpolation), parsing and slice generation done Charts done correcting peaks etc... various calculations ... fixed MZshift calculation changed RTshift calculation fixed peakPick beeing active when nextprev fixes to IOThread small fixes to peak picking Savitzky-Golay Peak picking Area Chart for Peaks peak index marker peak index marker for manual peaks peaks have to have a minimum length in both directions Re-storing slice information fixed ... gravity New shift calculation "gravity", visualization refinements on "gravity" ... fixed manual peak on stored slices gravity fitting various progressbars fixed fitting on batch shift is no longer displayed interpolated

+++ src/main/java/com/mycompany/fxmltableview/datamodel/Entry.java
@@ -585,8 +585,8 @@ if (listofSlices.containsKey(file)) {
     /**
      * @param fittedShift the OGroupfittedShift to set
      */
-    public void setFittedShift(RawDataFile file, short shift) {
-        this.Interpolatedshift.put(file,shift);
+    public void setFittedShift(RawDataFile file, float shift) {
+        this.Interpolatedshift.put(file,(short)shift);
         float shiftintime = session.getProparraycalculator().getshiftintime(shift);
         
         this.OgroupShift.put(file, shiftintime);
@@ -601,6 +601,8 @@ if (listofSlices.containsKey(file)) {
     }
         }
     }
+    
+     
 
 //    /**
 //     * @return the AdductPropArray
